FVC PL

+++ FVCalc.java
@@ -8,7 +8,7 @@ public class FVCalc {
 		Double rate = Double.parseDouble(args[1]);
 		int n = Integer.parseInt(args[2]);
 		double raten = (rate / 100);
-		double FutureValue = currentValue * (Math.pow(1 + raten, n));
+		double FutureValue = (int)(currentValue * (Math.pow(1 + raten, n)));
 		System.out.println("After " + n + " years" + "," + " a $" + currentValue + " saved at " + rate + "% "
 				+ "will yield $" + FutureValue);
